+++ algos_tf14/tf_moving_mean_std.py
