--- data.yml
@@ -1,58 +1,0 @@
-algo:
-  name: a2c_continuous
-config:
-  bounds_loss_coef: 0.5
-  clip_value: true
-  critic_coef: 1
-  e_clip: 0.2
-  entropy_coef: 0.0
-  env_name: BipedalWalkerHardcore-v2
-  gamma: 0.99
-  grad_norm: 0.5
-  learning_rate: 1e-4
-  lr_schedule: adaptive
-  lr_threshold: 0.008
-  max_epochs: 10000
-  mini_epochs: 4
-  minibatch_size: 2048
-  name: walker_0_0
-  normalize_advantage: true
-  normalize_input: false
-  num_actors: 16
-  ppo: true
-  reward_shaper:
-    scale_value: 0.1
-  score_to_win: 280
-  seq_len: 8
-  steps_num: 256
-  tau: 0.9
-  truncate_grads: true
-load_checkpoint: false
-load_path: path
-model:
-  name: continuous_a2c_logstd
-network:
-  mlp:
-    activation: relu
-    initializer:
-      name: normc_initializer
-      std: 1
-    regularizer:
-      name: None
-    units:
-    - 512
-    - 256
-    - 128
-  name: actor_critic
-  separate: true
-  space:
-    continuous:
-      fixed_sigma: true
-      mu_activation: None
-      mu_init:
-        name: normc_initializer
-        std: 0.01
-      sigma_activation: None
-      sigma_init:
-        name: const_initializer
-        value: 0
